+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 - Add rate limiting to the search and auth calls @robgietema
 - Add config option for request limits @robgietema
 - Added translations for 30 languages @mpalomaki @robgietema
+- Added scheduled task support @robgietema
 
 ### Bugfix